Added a callstream ( http://dbj.org/dbj/?p=514 ) interface for the do notation, which even provides a very convenient and efficient way to do self-recursion.
One way of doing that is to wrap a parser in function, which is easy but adds a considerable overhead because each parser will be recreated during every recursion.

Using late binding (i.e. adding an extra function before the parser) requires more typing, and it's easier to mistype, but it has no performance implications. This is also what the callstream syntax does behind the scenes.

Example:

var p = cs
 ("x", "<-", char_, "a")
 ("xs", "<-", optional, recurse)
 (ret, function(b){ return  b.xs ? [b.x, b.xs] : b.x });

run(p, "aaaaa");